Changed the criterion for a character to be a representative on computing line pitch

+++ pdf-models/src/main/java/model/PdfDocument.java
@@ -51,5 +51,5 @@ public interface PdfDocument extends HasDimensionStatistics, HasTextStatistics,
    */
   public void setSectionHeadingMarkup(String markup);  
   
-  public int getEstimatedLinePitch();
+  public float getEstimatedLinePitch();
 }
